Propose changes
Hi 
please I think the function in line 360 can cause some problems.
attributes: street, street_number and complement are not required because blank=True and null=True

and the function need to return something at the end of the day. Let suppose that all the three attribute are empty. that function will raise an error.

so the solution it's to remove at least null and blank to street attribute.