Add type stubs for DjangoRestFramework (#609)
* Add type stubs for django / mypy

* Fix env issues

And monkeypatch types

* Fix a variety of mypy errors

Removed useless warnings and added many missing type parameters. Also fixed a type mismatch in a variable assignment

* Update some deprecated or changed imports

u_gettext* is deprecated

* Add type parameters to many django types

* Use the actual status types instead of alias

* Fix OrganizationOwner serializer

It was based on the Person one, but the person serializer uses fields
that Organization does not have.

* Use StrOrPromise type

* Use asserts to satisfy mypy

Pretty sure the application would crash if you where logged out
when that code tries to run.

* Try to fix AuthUser type errors

Might be a better way. Also the AuthUser being based on AbstractBaseUser
instead af AbstractUser does make mypy complain, so I had to ignore a line

* Ignore an error and remove some ignores

Some ignore's where useless, the one I added is because I cant figure it out

* making flake8 happy

* Make pytype happy

It has trouble with model.TextChoices, indicating that it's also
an enum helps it.

* Fix translations

Docker was missing gettext and translations needed to be updated

* Add djangorestframework type stubs

* Add type parameters

* ignore some type errors

* fix translations

* Add type parameters

* Ignore some type errors

Cant figure out how to fix them right now

* Update translations

* Remove unused import

* Apply pr review changes

One of the issues could not be fixed (https://github.com/LesFruitsDefendus/saskatoon-ng/pull/608#discussion_r2442677794), but I made an issue about it and linked it in the code.