Add Equipment point reservation to harvest form (#612)
* Filter available equipment points

If any of it's  equipment is being used by a harvest during the required
period, that equipment point is unavailable

First step of implementing issue #420

* fix imports for harvest utils

* Expand EquipmentPointAutocomplete with filter

When start_date and end_date is forwarded, limit the possible points
to those that are not booked in that time range.

* Test the available_equipment_points function

Had to downgrade typeguard until deal can fix a breaking change in typeguard 4,
but all 7 tests are passing

* Fix flake warnings and make translations

* Tweak contracts for available_equipment_points

Added a contract that makes sure the user is using offset aware datetimes,
since the database one's are offset aware and you cant compare offset
naive datetime's with offset aware ones.

* fix typeguard version

deal has not been updated to use typeguard 4 yet

* fix overindented line

* available_equipment_point_reservation should only consider active harvests

* Improve available_equipment_points tests

Now we can test every status

* Add fixture for reservation conflict

Way in the future, should be a while before it's
no longer conflictable

* Test and typecheck member autocomplete classes

* Move available_equipment_points to member app

Since it's only used in member.autocomplete right now

* Make translations

* Typecheck HarvestForm

* Add equipment point reservation to harvest form

Mostly based on https://github.com/LesFruitsDefendus/saskatoon-ng/blob/7b399f3e495122900242c14e2c7b2af7460d8d86/saskatoon/harvest/forms.py

* add basic tests for HarvestForm

* remove autocomplete code from HarvestForm

it should be in the next pr

* Make Autocomplete class

Can be used instead of it's super to ease writting tests

* Improvements to equipment_points_available function and tests

* Add more info to valid date contract

* Make available_equipment_points tests more complete

by adding a second equipment to the equipment point and
testing all statuses

* Add comments to available_equipment_points

to explain the context behind how it works

* Fix a variety of review comments

* Add types and tests form HarvestForm

* format harvest.forms with yapf

had to ignore one flake rule to get it to work

* make translations

* Give a dedicated clean function to equipment in harvest form

* Allow available_equipment_points to consider the harvest itself in the reservation

also run yapf on utils and test_utils as well as add two tests to cover
the code changes.

* Simplify test_available_equipment_points_fuzz

* fix equipment points autocomplete

* Try to fix equipment reservation in harvest form

* Rework harvest form to remove hidden equipment_reserved field

* translations

* fix harvest form init test

* Get reserved equipment point from reserved equipment in harvest form

* try to add equipment to harvest detail view

* Put equipment point in harvest details view

* small improvement to serializers

* formatting pass

* update baseline data for serializer tests

* add typings for django-filter and type harvest class

* fix type error in RFC serializing

* make translations

* remove yapf

see https://github.com/LesFruitsDefendus/saskatoon-ng/pull/614

* fix post merge mypy errors

* fix pytype errors

* Had to simplify the type of rgetattr

* Add model tests and harvest reservation validation

We validate that the harvest reservation is still valid after
a status change. Also remove dependency on deprecated hypothesis[django]
as it was holding back the hypothesis version. This allowed me to finally
get the model tests working.

* Qynn wants keys to be single quote

* make translations

* Move get_equipment_point logic to serializer

and show only equipment point name and address in harvest details

* Add test for get_equipment_point

* fix test comment

* update migrations

since we changed the validation for some fields (to better inform hypothesis)

* fix form saving and improve harvest details

* fix bug with harvest creation

* make translations

* Add reservation details to harvest details

* improve equipment point display in harvest details

* fix quotes

* fix some problems with quotes

* exclude migrations from type checking

* try to fix some problems with indentation

* Fix a bunch of review comments

* Update saskatoon/harvest/models.py

Co-authored-by: Qynn Schwaab <41802320+2ynn@users.noreply.github.com>

* Update saskatoon/sitebase/templates/app/detail_views/harvest/view.html

Co-authored-by: Qynn Schwaab <41802320+2ynn@users.noreply.github.com>

* Fix a bunch of review comments

* qynn's nitpicks

* Add id check to autocomplete

it's useless to call the database if there's no id

* Remove initial id from harvest form

* improve equipment point autocomplete

* fix autocomplete crash when start > end

* harvest id should be a string

* fix harvest id parsing autocomplete

* fix some errors in harvest form

* fix datetime parsing

* more tests and stricter autocomplete

* if no date is given, send back all equipment points

* Add regression test for timezone handling in autocomplete

* fix nit

* fix harvest form population

* Fix problem with autocomplete

* Remove some useless logs and fix tests

* found a better way to do rgetattr

* Improvements to serializers

* better type ignore comments

* last pass on logging

---------

Co-authored-by: Qynn Schwaab <41802320+2ynn@users.noreply.github.com>